refactor(native): move runtime terrain into subdomain folder
