+++ src/native/engine/engine.c
@@ -32,9 +32,9 @@
 #include "runtime/input_abi.h"
 #include "runtime/resource_domain.h"
 #include "runtime/render_submit.h"
-#include "runtime/terrain_generation.h"
-#include "runtime/terrain_abi.h"
-#include "runtime/terrain_runtime.h"
+#include "runtime/terrain/terrain_generation.h"
+#include "runtime/terrain/terrain_abi.h"
+#include "runtime/terrain/terrain_runtime.h"
 #include "runtime/tick/tick_phases.h"
 #include "runtime/world/world_metrics.h"
 #include "runtime/world/world_abi.h"
