+++ src/native/engine/runtime/application_service_ports.c
@@ -4,9 +4,9 @@
 #include "player/player_runtime_service.h"
 #include "render_material.h"
 #include "render_submit.h"
-#include "terrain_height.h"
-#include "terrain_host.h"
-#include "terrain_mutation.h"
+#include "terrain/terrain_height.h"
+#include "terrain/terrain_host.h"
+#include "terrain/terrain_mutation.h"
 
 #include <string.h>
