+++ src/native/engine/runtime/terrain/terrain_abi.c
