+++ src/native/engine/runtime/terrain/terrain_height.c
