+++ src/native/engine/runtime/terrain/terrain_host.c
