+++ src/native/scaffold/native_entry.c
@@ -1,6 +1,6 @@
 #include "banana_native_v3.h"
 #include "runtime/chunk_stream_packet.h"
-#include "runtime/terrain_chunks.h"
+#include "runtime/terrain/terrain_chunks.h"
 
 #include <stdio.h>
 #include <stdint.h>
